Theme extractor job done
