+++ test/jobs/theme_extractor_job_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class ThemeExtractorJobTest < ActiveJob::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
